% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capacity_plot.R
\name{get_capacity_plot}
\alias{get_capacity_plot}
\title{Generate Capacity Plot for Sailings}
\usage{
get_capacity_plot(sailings_data)
}
\arguments{
\item{sailings_data}{A dataframe containing at least four columns: `sailingStatus`, `time`,
`carFill`, and `oversizeFill`. `sailingStatus` should indicate whether
a sailing is current or future, `time` should represent the time of each
sailing, and `carFill` and `oversizeFill`should contain fill percentages
for cars and oversize vehicles, respectively.}
}
\value{
A ggplot object representing the fill percentages of vehicles for current
and future sailings. The plot is color-coded to distinguish between car and
oversize vehicle fill percentages.
}
\description{
This function takes a dataset of sailings and generates a bar plot showing
the fill percentages for cars and oversize vehicles for current and future
sailings, distinguished by time. The plot uses different colors to represent
different vehicle types and displays the fill percentage across various times
of the day.
}
\examples{
sallings_data = get_capacity("TSASWB")
get_capacity_plot(sallings_data)
}
