% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capacity.R, R/get_capacity_plot.R,
%   R/get_schedule.R
\name{get_combined_sailings_df}
\alias{get_combined_sailings_df}
\title{Combine Sailings Data with Route Information}
\usage{
get_combined_sailings_df()

get_combined_sailings_df()

get_combined_sailings_df()
}
\value{
A data frame that combines sailings data with route information. Each row
represents a specific sailing, including details such as route code, from
and to terminal codes, sailing duration, time, arrival time, sailing status,
vessel name, fill percentages for cars and oversize vehicles.

A data frame that combines sailings data with route information. Each row
represents a specific sailing, including details such as route code, from
and to terminal codes, sailing duration, time, arrival time, sailing status,
vessel name, fill percentages for cars and oversize vehicles.

A data frame that combines sailings data with route information. Each row
represents a specific sailing, including details such as route code, from
and to terminal codes, sailing duration, time, arrival time, sailing status,
vessel name, fill percentages for cars and oversize vehicles.
}
\description{
This function fetches data from the BC Ferries API and processes it to
combine sailings data with corresponding route information. It extracts
specific details such as route codes, terminal codes,sailing duration,
and other sailings information like time, arrival time, sailing status,
and vessel name.

This function fetches data from the BC Ferries API and processes it to
combine sailings data with corresponding route information. It extracts
specific details such as route codes, terminal codes,sailing duration,
and other sailings information like time, arrival time, sailing status,
and vessel name.

This function fetches data from the BC Ferries API and processes it to
combine sailings data with corresponding route information. It extracts
specific details such as route codes, terminal codes,sailing duration,
and other sailings information like time, arrival time, sailing status,
and vessel name.
}
\examples{
combined_sailings_df <- get_combined_sailings_df()
head(combined_sailings_df)
combined_sailings_df <- get_combined_sailings_df()
head(combined_sailings_df)
combined_sailings_df <- get_combined_sailings_df()
head(combined_sailings_df)
}
