% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capacity.R, R/get_capacity_plot.R,
%   R/get_routes.R, R/get_schedule.R
\name{get_data}
\alias{get_data}
\title{Fetch Ferries Data from BC Ferries API}
\usage{
get_data()

get_data()

get_data()

get_data()
}
\value{
A list or data frame containing the ferries data, if the request is
successful. If the request fails, the function stops with an error message.

A list or data frame containing the ferries data, if the request is
successful. If the request fails, the function stops with an error message.

A list or data frame containing the ferries data, if the request is
successful. If the request fails, the function stops with an error message.

A list or data frame containing the ferries data, if the request is
successful. If the request fails, the function stops with an error message.
}
\description{
This function is designed to interact with the BC Ferries API,
specifically targeting the capacity endpoint. It sends a GET request to the
API and if successful, parses the returned JSON content into an R-readable
format. The function is robust against failed requests, providing an error
message if the data retrieval is unsuccessful.

This function is designed to interact with the BC Ferries API,
specifically targeting the capacity endpoint. It sends a GET request to the
API and if successful, parses the returned JSON content into an R-readable
format. The function is robust against failed requests, providing an error
message if the data retrieval is unsuccessful.

This function is designed to interact with the BC Ferries API,
specifically targeting the capacity endpoint. It sends a GET request to the
API and if successful, parses the returned JSON content into an R-readable
format. The function is robust against failed requests, providing an error
message if the data retrieval is unsuccessful.

This function is designed to interact with the BC Ferries API,
specifically targeting the capacity endpoint. It sends a GET request to the
API and if successful, parses the returned JSON content into an R-readable
format. The function is robust against failed requests, providing an error
message if the data retrieval is unsuccessful.
}
\examples{
api_data = get_data()
print(api_data)
api_data = get_data()
print(api_data)
api_data = get_data()
print(api_data)
api_data = get_data()
print(api_data)
}
