% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes.R
\name{get_routes}
\alias{get_routes}
\title{Showing Ferry Route Information}
\usage{
get_routes()
}
\value{
A df contains ferry routes information: routeCode, fromTerminalCode
and to TerminalCode.
}
\description{
This function obtains ferry route information.
}
\examples{
route_info <- get_routes()
print(route_info)
}
